Create `ReviewSummary` component
